Moved entities and used shared context
