+++ src/Entities/Data/Analytics/DirectoryAnalyticMetric.cs
@@ -1,10 +1,6 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
-namespace Analytics.Data.Entities
+namespace Biobanks.Entities.Data.Analytics
 {
     public class DirectoryAnalyticMetric
     {
